requirements.txt
flask>=2.2.5
sentence-transformers>=2.2.2
chromadb>=0.4.15
python-dotenv>=1.0.0
google-generativeai>=0.3.2
requests>=2.31.0
beautifulsoup4>=4.12.2